// projects/staff/src/app/api/main/v1/model/ticket.ts
/**
 * UG Examine
 * UG Examine API description
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { User } from './user';


export interface Ticket { 
    id: string;
    description?: string;
    studentID: string;
    status: Ticket.StatusEnum;
    course: string;
    year: string;
    semmester: Ticket.SemmesterEnum;
    reporter: User;
    assignee: User;
}
export namespace Ticket {
    export type StatusEnum = 'created' | 'open' | 'investigating' | 'closed';
    export const StatusEnum = {
        Created: 'created' as StatusEnum,
        Open: 'open' as StatusEnum,
        Investigating: 'investigating' as StatusEnum,
        Closed: 'closed' as StatusEnum
    };
    export type SemmesterEnum = 1 | 2;
    export const SemmesterEnum = {
        NUMBER_1: 1 as SemmesterEnum,
        NUMBER_2: 2 as SemmesterEnum
    };
}